+++ Assets/Animations/DocAnim.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1153ee2588b2882439bc481251698402
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
