+++ Assets/Animations/idle_down.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2e148302b882e5b41abca90d2b2eba2a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
