+++ Assets/Animations/idle_left.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 01c13a12b7d0d0f42845f4bf6241aedc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
