+++ Assets/Animations/idle_right.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7a197939a7bae8240b544639080d7d8f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
