+++ Assets/Animations/idle_up.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 202e93772fd9f0c479091180650a13ae
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
