+++ Assets/Animations/walk_down.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 515b888bc1b3f424f8ed6320d6f2e2ae
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
