+++ Assets/Animations/walk_left.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eb8c1278b466f854cbcfeb74150f6f42
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
