+++ Assets/Animations/walk_right.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b47d7cc67296fec4099c4e95fde46936
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
