+++ Assets/Animations/walk_up.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ba8c0ebcd1e48a2449a3b6ddb212779a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
