+++ Assets/Prefabs/Bullet.prefab
@@ -134,4 +134,4 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   speed: 10
-  lifetime: 3.5
+  lifetime: 1
