+++ Assets/Scripts/Bullet.cs
@@ -16,11 +16,13 @@ public class Bullet : MonoBehaviour
         transform.position += transform.up * speed * Time.deltaTime;
     }
 
-private void OnTriggerEnter2D(Collider2D collision)
+    private void OnTriggerEnter2D(Collider2D collision)
     {
-        if (collision.CompareTag("Enemy"))
+        if (collision.CompareTag("Enemy") || (collision.CompareTag("Wall")))
         {
             Destroy(gameObject);
         }
+
     }
 }
+
