+++ Assets/Scripts/aim.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2b82ffdd41f070249acf7b80bc989399
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
