+++ Assets/Sprites/Doc.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2a96894afd314f944b48c58ff2eec478
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
