+++ ProjectSettings/TagManager.asset
@@ -9,6 +9,7 @@ TagManager:
   - Enemy
   - Teleporter1
   - Teleporter2
+  - Wall
   layers:
   - Default
   - TransparentFX
